pyproject.toml
[project]
name = "stepik-autotest"
version = "0.1.0"
description = ""
authors = [
    {name = "artur.balsis",email = "artur.balsis@gmail.com"}
]
readme = "README.md"
requires-python = "^3.11"
dependencies = [
    "pytest==8.3.5",
    "selene==2.0.0rc9",
    "selenium~=4.30.0",
    "allure-pytest>=2.13.5",
    "python-dotenv>=1.1.0",
    "pytest-rerunfailures==14.0",
    "pytest-xdist==3.6.1",
    "pytils==0.4.1",
    "requests (>=2.32.3,<3.0.0)",
    "appium-python-client (>=5.0.0,<6.0.0)",
    "faker (>=37.1.0,<38.0.0)",
    "pydantic (>=2.11.3,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
pylint = "^3.3.6"

